added user service. fixed good factor calculating. start of jwt working

+++ GoodNewsAggregator/GoodNewsAggregator.Core.DTO/ArticleWithRssNameDto.cs
@@ -13,7 +13,7 @@ namespace GoodNewsAggregator.Core.DTO
         public string Title { get; set; }
         public string Content { get; set; }
         public DateTime Date { get; set; }
-        public int GoodFactor { get; set; }
+        public float GoodFactor { get; set; }
 
         public Guid RssId { get; set; }
         public string RssName { get; set; }
